# internal/routes/helthCheck.go
// Package routes provides HTTP route handlers including health checks.
package routes

import (
	//"fmt"
	"github.com/gin-gonic/gin"
	"net/http"
	"time"
)

type ResponseHealthCheck struct {
	Status      string `json:"Status"`
	RecvTime    string `json:"recv_time"`
	RecvTimeUTC string `json:"recv_time_utc"`
}

// HealthCheck API
// @Summary Check health status
// @Description Returns OK if the service is healthy
// @Tags health
// @Success 200 {object} ResponseHealthCheck "Health check response"
// @Router /healthz [get]
func HealthCheck(c *gin.Context) {
	c.Header("Content-Type", "application/json")
	response := ResponseHealthCheck{
		Status:      "OK",
		RecvTime:    time.Now().Format("2006-01-02T15:04:05"),
		RecvTimeUTC: time.Now().UTC().Format(time.RFC3339),
	}
	c.JSON(http.StatusOK, response)
}